Move profiler.resolveAllSources to compile-common
- Abstract getImports as function parameter instead of requiring
  solc and parserSolc

- Retain profiler.resolveAllSources as wrapper to construct getImports()
