+++ packages/compile-common/src/profiler/index.ts
@@ -4,3 +4,4 @@ export { readAndParseArtifactFiles } from "./readAndParseArtifactFiles";
 export { findUpdatedFiles } from "./findUpdatedFiles";
 export { isExplicitlyRelative } from "./isExplicitlyRelative";
 export { minimumUpdatedTimePerSource } from "./minimumUpdatedTimePerSource";
+export { resolveAllSources } from "./resolveAllSources";
